Refactor NetworkManager using singleton
